LMAO
it's LMAO bby

LMAO is a pair of strings. one of them is used as a tag while the other is associated data. two LMAOs with equal tags are are "equal" (more accurately equivalent) even if the data are different.

this is because so much of the 4,132 extra matches are not actually extra matches. many of them are equivalent to one another but where the only thing different is whitespace.