prep for string literals
theory. in translation phase 6 string literals get concatenated into one and in phase 7 '\0' is appended.

plan. using a Builder we create a LazyByteString that holds a reference to spans of the text. when a chunk of the LazyByteString is forced, that part of the input will be copied into memory.