fix function parsing bug; reduce backtracking when identifier is optional in declarator parsing
function parsing bug:

something like

static int g();

was being parsed as an ordinarily specified function 'g' returning a 'static int' value, which is nonsense.

the correct interpretation must be: 'static'-specified 'g' returning a (regular) 'int'.

issue was a byproduct of our type parser design that combines storage class specifiers with other qualifiers. tradeoffs between simplicity and ontological cleanness.

fourtunately fix is pretty easy. just manually steal the specifiers from the return type and move them to the function type itself.

backtracking issue when parsing a declarator:

previously we had two declarator parsers: one for regular declarators, which expect identifiers to be declared, and abstract declarators, which do not. in short, the difference is between:

- x[3], and
- [3].

abstract declarators are used for naming types and in function parameter lists.

a function parameter list can expect a mixture of (regular) and abstract declarators, which in our design requires backtracking. fortunately, it's pretty easy to remove the backtracking, by just adding a third option when parsing.

here a slight optimization was applied where instead of passing in an enumeration to decide whether to parse an identifier, i just pass in a function that decides what to do. this saves branching by streamlining the parser. declarators come often but the performance impact is unknown.